refactor(a2a): extract A2AService with prepare/run split (PR #69 pattern)
Extracts a dedicated A2AService with:
- send_message / get_task / cancel_task / agent_def_to_card
- prepare_streaming_session / run_streaming_session
- prepare_resubscribe / run_resubscribe_session
- handle() as the top-level JSON-RPC dispatcher

All fallible work now runs in prepare_* methods and returns typed
StreamingSession / ResubscribeSession structs. run_* methods are
infallible — pure event->SSE frame mapping with terminal handling.
A2AHandler becomes a thin adapter that constructs an A2AService.

Runtime invariants preserved:
- Only orchestrator.register_task + spawn_task_relay +
  spawn_background_execution for task mechanics. No direct
  broadcaster.publish or coordinator mutations from the service.
- Preflight validate_provider_secrets runs in caller's task-local
  context (user/workspace from middleware), preserving TenantSecretStore
  access.
- Zero hand-rolled JsonRpcResponse construction literals (from Phase 5a).

Preserves the public A2AHandler surface so existing callers compile.
