+++ server/distri-core/src/a2a/stream.rs
@@ -1,25 +1,23 @@
-use crate::a2a::handler::validate_message;
-use crate::a2a::mapper::{map_agent_event, map_final_result};
-use crate::a2a::{extract_text_from_message, SseMessage};
-use crate::agent::InvokeResult;
+//! Preflight helpers for A2A streaming: provider-secret validation, thread
+//! init, execution-context preparation, background execution spawn.
+//!
+//! The top-level streaming entry points live in `a2a/service.rs` (the
+//! `A2AService::prepare_streaming_session` + `run_streaming_session` pair).
+//! Everything here is re-used from there.
+
+use crate::a2a::extract_text_from_message;
+use crate::a2a::validate_message;
 use crate::agent::{
     types::ExecutorContextMetadata, AgentEventType, AgentOrchestrator, ExecutorContext,
 };
 use crate::secrets::SecretResolver;
 use crate::AgentError;
-use distri_auth::context::{with_user_and_workspace, with_user_id};
-// Note: with_user_and_workspace IS needed for stream! macro and spawned tasks
-// because they don't inherit task-local storage from middleware
 use anyhow::anyhow;
-use distri_a2a::{JsonRpcError, MessageSendParams};
+use distri_a2a::MessageSendParams;
+use distri_auth::context::with_user_and_workspace;
 use distri_types::configuration::{AgentConfig, DefinitionOverrides};
 
-use futures_util::future::poll_fn;
-use futures_util::Stream;
-use std::future::Future;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 
 /// Validates that required provider secrets are configured before execution starts.
 /// This provides an early, user-friendly error message instead of failing mid-stream.
@@ -58,35 +56,6 @@ pub async fn validate_provider_secrets(
     }
 
     Ok(())
-}
-
-struct UserScopedStream<S> {
-    user_id: String,
-    inner: Pin<Box<S>>,
-}
-
-impl<S> UserScopedStream<S> {
-    fn new(user_id: String, inner: Pin<Box<S>>) -> Self {
-        Self { user_id, inner }
-    }
-}
-
-impl<S> Stream for UserScopedStream<S>
-where
-    S: Stream,
-{
-    type Item = S::Item;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-        let user_id = this.user_id.clone();
-        let mut inner = this.inner.as_mut();
-        let mut fut = Box::pin(with_user_id(
-            user_id,
-            poll_fn(|cx| inner.as_mut().poll_next(cx)),
-        ));
-        fut.as_mut().poll(cx)
-    }
 }
 
 /// Create browser session via browsr and return session info.
@@ -154,40 +123,9 @@ pub async fn init_thread_get_message(
     Ok((thread_id, message))
 }
 
-/// Subscribe to events for an existing task via the broadcaster.
-/// Used by the `tasks/resubscribe` A2A method.
-pub async fn handle_resubscribe_sse(
-    req_id: Option<serde_json::Value>,
-    task_id: String,
-    executor: Arc<AgentOrchestrator>,
-) -> impl futures_util::stream::Stream<Item = Result<SseMessage, std::convert::Infallible>> {
-    async_stream::stream! {
-        // Subscribe via broadcaster (always available via runtime)
-        let event_stream = match executor.broadcaster().subscribe(&task_id).await {
-            Ok(s) => s,
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    req_id.clone(),
-                    JsonRpcError::internal(format!("Failed to subscribe: {}", e)),
-                ));
-                return;
-            }
-        };
-
-        futures_util::pin_mut!(event_stream);
-        while let Some(event) = futures_util::StreamExt::next(&mut event_stream).await {
-            let msg = map_agent_event(&event);
-            yield Ok::<_, std::convert::Infallible>(SseMessage::success_frame(
-                req_id.clone(),
-                serde_json::to_value(&msg).unwrap_or_default(),
-            ));
-        }
-    }
-}
-
 /// Prepare ExecutorContext with metadata, browser session, definition overrides, etc.
-/// Returns (ExecutorContext, message, definition_overrides) or an error.
-async fn prepare_execution(
+/// Returns `(ExecutorContext, definition_overrides)` or an error.
+pub async fn prepare_execution(
     agent_id: &str,
     params: &MessageSendParams,
     executor: &Arc<AgentOrchestrator>,
@@ -338,182 +276,4 @@ pub(crate) fn spawn_background_execution(
             }
         }
     }));
-}
-
-pub async fn handle_message_send_streaming_sse(
-    req_id: Option<serde_json::Value>,
-    agent_id: String,
-    params: serde_json::Value,
-    executor: Arc<AgentOrchestrator>,
-    executor_context: Arc<ExecutorContext>,
-) -> impl futures_util::stream::Stream<Item = Result<SseMessage, std::convert::Infallible>> {
-    let user_id = executor_context.user_id.clone();
-    let stream_user_id = user_id.clone();
-    let workspace_id = executor_context
-        .workspace_id
-        .as_ref()
-        .and_then(|s| uuid::Uuid::parse_str(s).ok());
-    let stream_workspace_id = workspace_id;
-    let id_field_clone = executor_context.session_id.clone();
-
-    // Validate provider secrets BEFORE entering the stream! macro,
-    // because stream! doesn't inherit task-local storage (user/workspace context)
-    // needed by TenantSecretStore.
-    let secret_validation = validate_provider_secrets(&executor, &agent_id).await;
-
-    let stream = async_stream::stream! {
-        let user_id = stream_user_id.clone();
-        let params: MessageSendParams = match serde_json::from_value(params) {
-            Ok(p) => p,
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    Some(id_field_clone.clone().into()),
-                    JsonRpcError::invalid_params(e.to_string()),
-                ));
-                return;
-            }
-        };
-
-        // Check the pre-computed secret validation result
-        if let Err(e) = secret_validation {
-            yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                Some(id_field_clone.clone().into()),
-                JsonRpcError::internal(e.to_string()),
-            ));
-            return;
-        }
-
-        // stream! macro doesn't inherit task-local storage, so wrap here
-        let (thread_id, message) = match with_user_and_workspace(
-            user_id.clone(),
-            stream_workspace_id,
-            init_thread_get_message(
-                agent_id.clone(),
-                executor.clone(),
-                &params,
-                executor_context.clone(),
-            )
-        )
-        .await
-        {
-            Ok(t) => t,
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    Some(id_field_clone.clone().into()),
-                    JsonRpcError::internal(e.to_string()),
-                ));
-                return;
-            }
-        };
-
-        // Prepare execution context with metadata, browser, overrides
-        let (exec_ctx, definition_overrides) = match prepare_execution(
-            &agent_id, &params, &executor, &executor_context,
-        ).await {
-            Ok(v) => v,
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    Some(id_field_clone.clone().into()),
-                    JsonRpcError::internal(e.to_string()),
-                ));
-                return;
-            }
-        };
-
-        let main_task_id = exec_ctx.task_id.clone();
-
-        // === Background execution: register → relay → spawn → subscribe ===
-        // Client disconnect does NOT kill the agent — execution continues in background.
-
-        // 1. Register task — wire cancellation signal + mailbox into context
-        let (executor_context_arc, event_rx) = match executor
-            .register_task(&main_task_id, &thread_id, exec_ctx)
-            .await
-        {
-            Ok(v) => v,
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    req_id.clone(),
-                    JsonRpcError::internal(format!("Failed to register task: {}", e)),
-                ));
-                return;
-            }
-        };
-
-        // 2. Spawn event relay: agent events → broadcaster
-        executor.spawn_task_relay(main_task_id.clone(), event_rx);
-
-        // Clone *before* spawn_background_execution moves the Arc — this is
-        // the context the `final` tool writes its result to, so we must read
-        // from the same Arc after the agent completes.
-        let executor_context_for_final = executor_context_arc.clone();
-
-        // 3. Spawn background execution
-        spawn_background_execution(
-            executor.clone(),
-            agent_id.clone(),
-            message,
-            executor_context_arc,
-            Some(definition_overrides),
-            main_task_id.clone(),
-            user_id.clone(),
-            workspace_id,
-        );
-        match executor.broadcaster().subscribe(&main_task_id).await {
-            Ok(event_stream) => {
-                futures_util::pin_mut!(event_stream);
-                let mut saw_terminal = false;
-                while let Some(event) = futures_util::StreamExt::next(&mut event_stream).await {
-                    let is_terminal = matches!(
-                        &event.event,
-                        AgentEventType::RunFinished { .. } | AgentEventType::RunError { .. }
-                    );
-                    let msg = map_agent_event(&event);
-                    yield Ok::<_, std::convert::Infallible>(SseMessage::success_frame(
-                        req_id.clone(),
-                        serde_json::to_value(msg).unwrap_or_default(),
-                    ));
-                    if is_terminal {
-                        saw_terminal = true;
-                        break;
-                    }
-                }
-
-                // After terminal event: read the final result from the shared
-                // ExecutorContext (set by the `final` tool via set_final_result, and
-                // shared via Arc<RwLock>) and yield it as MessageKind::Message so
-                // clients render the final answer.
-                if saw_terminal {
-                    if let Some(final_value) =
-                        executor_context_for_final.get_final_result().await
-                    {
-                        let text = match final_value {
-                            serde_json::Value::String(s) => s,
-                            other => other.to_string(),
-                        };
-                        if !text.is_empty() {
-                            let result = InvokeResult {
-                                content: Some(text),
-                                ..Default::default()
-                            };
-                            let msg = map_final_result(&result, executor_context_for_final);
-                            yield Ok::<_, std::convert::Infallible>(SseMessage::success_frame(
-                                req_id.clone(),
-                                serde_json::to_value(msg).unwrap_or_default(),
-                            ));
-                        }
-                    }
-                }
-            }
-            Err(e) => {
-                yield Ok::<_, std::convert::Infallible>(SseMessage::error_frame(
-                    req_id.clone(),
-                    JsonRpcError::internal(format!("Failed to subscribe to task events: {}", e)),
-                ));
-            }
-        }
-
-    };
-
-    UserScopedStream::new(user_id, Box::pin(stream))
 }
